Fix clang -Wconversion warnings
* include/wget/wget.h (wget_ready_2_transfer): Change 'mode' to int
* libwget/io.c: Likewise
* libwget/hpkp.c: Explicit int to bit conversion
* libwget/http.c: Likewise
* libwget/net.c: Likewise
* libwget/ocsp.c: Likewise
* libwget/ssl_gnutls.c: Likewise
* tests/libtest.c: Likewise
* libwget/iri.c (wget_iri_unescape_inline): Explicit int to char conversion
* libwget/utils.c: Likewise

+++ include/wget/wget.h
@@ -286,7 +286,7 @@ WGETAPI int
 WGETAPI int
 	wget_ready_2_write(int fd, int timeout);
 WGETAPI int
-	wget_ready_2_transfer(int fd, int timeout, short mode);
+	wget_ready_2_transfer(int fd, int timeout, int mode);
 WGETAPI int
 	wget_strcmp(const char *s1, const char *s2) G_GNUC_WGET_PURE;
 WGETAPI int
